Refactor DTOs location
