+++ src/main/java/com/example/fabrickcontroller/service/TransactionService.java
@@ -1,6 +1,6 @@
 package com.example.fabrickcontroller.service;
 
-import com.example.fabrickcontroller.domain.MoneyTransferDomainDto;
+import com.example.fabrickcontroller.dto.MoneyTransferDomainDto;
 import org.jetbrains.annotations.NotNull;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
